# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Modular E-commerce API",
        "description": "This is the API documentation for the Laravel Modular E-commerce API.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API server"
        }
    ],
    "paths": {
        "/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get list of categories",
                "operationId": "22f469a44156ddb19cb2b2d57fbda8fc",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "en",
                                "ar"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of categories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Categories retrieved successfully."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Category"
                                            }
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "status": true,
                                    "message": "Categories retrieved successfully.",
                                    "data": [
                                        {
                                            "id": 1,
                                            "name": "Electronics",
                                            "created_at": "2025-08-11T10:00:00Z",
                                            "updated_at": "2025-08-11T10:00:00Z"
                                        },
                                        {
                                            "id": 2,
                                            "name": "Books",
                                            "created_at": "2025-08-11T10:05:00Z",
                                            "updated_at": "2025-08-11T10:05:00Z"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Create a new category",
                "description": "Create a category with English and Arabic names and an image.",
                "operationId": "732d42dba66b1b750d27436a2da1d478",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "properties": {
                                            "en": {
                                                "type": "string",
                                                "example": "Electronics"
                                            },
                                            "ar": {
                                                "type": "string",
                                                "example": "إلكترونيات"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "image": {
                                        "type": "file",
                                        "format": "binary",
                                        "example": "category.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Category created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "properties": {
                                                "en": {
                                                    "type": "string",
                                                    "example": "Electronics"
                                                },
                                                "ar": {
                                                    "type": "string",
                                                    "example": "إلكترونيات"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "image": {
                                            "type": "string",
                                            "example": "https://example.com/images/category.jpg"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-08-11T12:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-08-11T12:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get category details by ID",
                "description": "Retrieve a single category by its unique ID.",
                "operationId": "ee15b3756492d3964803490b0337213d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the category",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 6
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "تم استرجاع الفئة بنجاح."
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "id": 6,
                                                "name": "إلكترونيات",
                                                "image": "http://localhost:8000/uploads/categories/categories/electronics.jpg",
                                                "created_at": "2025-08-11 18:04:38",
                                                "updated_at": "2025-08-11 18:04:38"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "الفئة غير موجودة."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            },
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/categories/{id}": {
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Update a category",
                "operationId": "d7826d129cf19525e77a16bd3a714705",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "properties": {
                                            "en": {
                                                "type": "string",
                                                "example": "Toys"
                                            },
                                            "ar": {
                                                "type": "string",
                                                "example": "ألعاب"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "image": {
                                        "description": "Image file or URL",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Category updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Category updated successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 26
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Toys"
                                                },
                                                "image": {
                                                    "type": "string",
                                                    "example": "http://localhost:8000/uploads/categories/Image_c66d33aa-4655-46eb-a7b0-8cdb08572c2b_1754934136.png"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2025-08-11 17:35:02"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2025-08-11 17:42:16"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Delete a category",
                "operationId": "31bf0686d13dbdb052c4ae2d54200f6b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Category deleted successfully."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Category": {
                "title": "Category",
                "description": "Category model",
                "properties": {
                    "id": {
                        "description": "Category ID",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Category name",
                        "type": "string"
                    },
                    "image": {
                        "description": "Category image URL",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Created timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Updated timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}